extra module type into own file
